removed circular dependency: separate the auth package into different files
